Update transaction, revocation at state update
Finished(I think?) the update transaction

Every time a state update happens the node checks whether a revocation needs to happen about a prefix it had previously assigned.

Managed to solve the problem with the broadcast_resolve_message() with request_futures (apparently async is now an independent module)
but... I have another problem with blocking requests to be discussed in our next meeting! :)